Update WebApiOutputCacheAttribute.cs
I have found another bug.  An exception occurs when no Accept headers are passed in with the request.

This causes a NullReferenceException.

I have changed it to use "Default" if no Accept header has been passed in.  

You may well have a better way of achieving this, so please refactor as you see fit.