Updated Comparison of SupportedMediaTypes to use the MediaType description so it can support parameters
This is to support custom media types that contain parameters. For
example:
application/vnd.mine+json
and
application/vnd.mine+json; version 2

those are effectively the same media types but the second contain a
'version' parameter. The existing comparison will not match it with the
Supported ones but the updated version will.